Pushed by SAMANTHA!

+++ core/src/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/Burning.java
@@ -1,4 +1,23 @@
-
+/*
+ * Pixel Dungeon
+ * Copyright (C) 2012-2015  Oleg Dolya
+ *
+ * Shattered Pixel Dungeon
+ * Copyright (C) 2014-2016 Evan Debenham
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>
+ */
 package com.shatteredpixel.shatteredpixeldungeon.actors.buffs;
 
 import com.shatteredpixel.shatteredpixeldungeon.Badges;
@@ -8,13 +27,17 @@ import com.shatteredpixel.shatteredpixeldungeon.actors.blobs.Blob;
 import com.shatteredpixel.shatteredpixeldungeon.actors.blobs.Fire;
 import com.shatteredpixel.shatteredpixeldungeon.actors.hero.Hero;
 import com.shatteredpixel.shatteredpixeldungeon.actors.mobs.Thief;
+import com.shatteredpixel.shatteredpixeldungeon.effects.Speck;
 import com.shatteredpixel.shatteredpixeldungeon.effects.particles.ElmoParticle;
 import com.shatteredpixel.shatteredpixeldungeon.items.Heap;
 import com.shatteredpixel.shatteredpixeldungeon.items.Item;
+import com.shatteredpixel.shatteredpixeldungeon.items.armor.glyphs.Brimstone;
 import com.shatteredpixel.shatteredpixeldungeon.items.food.ChargrilledMeat;
 import com.shatteredpixel.shatteredpixeldungeon.items.food.MysteryMeat;
 import com.shatteredpixel.shatteredpixeldungeon.items.rings.RingOfElements.Resistance;
 import com.shatteredpixel.shatteredpixeldungeon.items.scrolls.Scroll;
+import com.shatteredpixel.shatteredpixeldungeon.items.scrolls.ScrollOfMagicalInfusion;
+import com.shatteredpixel.shatteredpixeldungeon.items.scrolls.ScrollOfUpgrade;
 import com.shatteredpixel.shatteredpixeldungeon.levels.Level;
 import com.shatteredpixel.shatteredpixeldungeon.messages.Messages;
 import com.shatteredpixel.shatteredpixeldungeon.scenes.GameScene;
@@ -25,8 +48,6 @@ import com.watabou.utils.Bundle;
 import com.watabou.utils.Random;
 
 public class Burning extends Buff implements Hero.Doom {
-
-	private static final String TXT_BURNS_UP		= "%s burns up!";
 	
 	private static final float DURATION = 8f;
 	
@@ -57,37 +78,55 @@ public class Burning extends Buff implements Hero.Doom {
 
 			//maximum damage scales from 6 to 2 depending on remaining hp.
 			int maxDmg = 3 + Math.round( 4 * target.HP / (float)target.HT );
-			target.damage( Random.Int( 1, maxDmg ), this );
+			int damage = Random.Int( 1, maxDmg );
 			Buff.detach( target, Chill.class);
 
 			if (target instanceof Hero) {
 
 				Hero hero = (Hero)target;
-				Item item = hero.belongings.randomUnequipped();
-				if (item instanceof Scroll) {
-					
-					item = item.detach( hero.belongings.backpack );
-					GLog.w( Messages.get(this, "burnsup", item.toString()) );
-					
-					Heap.burnFX( hero.pos );
-					
-				} else if (item instanceof MysteryMeat) {
-					
-					item = item.detach( hero.belongings.backpack );
-					ChargrilledMeat steak = new ChargrilledMeat();
-					if (!steak.collect( hero.belongings.backpack )) {
-						Dungeon.level.drop( steak, hero.pos ).sprite.drop();
+
+				if (hero.belongings.armor != null && hero.belongings.armor.hasGlyph(Brimstone.class)){
+
+					int heal = hero.belongings.armor.level()/2;
+					if (heal > 0 && hero.HP < hero.HT) {
+						hero.sprite.emitter().burst(Speck.factory(Speck.HEALING), 1);
+						hero.HP = Math.min(hero.HT, hero.HP + heal);
 					}
-					GLog.w( Messages.get(this, "burnsup", item.toString()) );
-					
-					Heap.burnFX( hero.pos );
-					
+
+				} else {
+
+					hero.damage( damage, this );
+					Item item = hero.belongings.randomUnequipped();
+					if (item instanceof Scroll
+							&& !(item instanceof ScrollOfUpgrade || item instanceof ScrollOfMagicalInfusion)) {
+
+						item = item.detach( hero.belongings.backpack );
+						GLog.w( Messages.get(this, "burnsup", Messages.capitalize(item.toString())) );
+
+						Heap.burnFX( hero.pos );
+
+					} else if (item instanceof MysteryMeat) {
+
+						item = item.detach( hero.belongings.backpack );
+						ChargrilledMeat steak = new ChargrilledMeat();
+						if (!steak.collect( hero.belongings.backpack )) {
+							Dungeon.level.drop( steak, hero.pos ).sprite.drop();
+						}
+						GLog.w( Messages.get(this, "burnsup", item.toString()) );
+
+						Heap.burnFX( hero.pos );
+
+					}
+
 				}
 				
-			} else if (target instanceof Thief && ((Thief)target).item instanceof Scroll) {
-				
-				((Thief)target).item = null;
-				target.sprite.emitter().burst( ElmoParticle.FACTORY, 6 );
+			} else {
+				target.damage( damage, this );
+			}
+
+			if (target instanceof Thief && ((Thief)target).item instanceof Scroll) {
+					((Thief)target).item = null;
+					target.sprite.emitter().burst( ElmoParticle.FACTORY, 6 );
 			}
 
 		} else {
